+++ blocks/header/header.css
@@ -23,6 +23,7 @@ header {
     height: 32px !important;
   }
 }
+
 /* header css end */
