+++ scripts/common.js
@@ -1,3 +1,4 @@
+import { fetchPlaceholders, getMetadata } from './aem.js';
 import Category from './category.js';
 import Tag from './tag.js';
 
@@ -49,4 +50,61 @@ export async function getList() {
   // console.log(list);
   proxy.list = list;
   return proxy.list;
+}
+
+async function buildBreadcrumbsFromNavTree() {
+  const crumbs = [];
+
+  // const homeUrl = document.querySelector('.nav-brand a[href]').href;
+  const currentUrl = new URL(window.location.href);
+  const currentPath = currentUrl.pathname;
+  const queryList = await getQueryList();
+  const breadcrumbsList = queryList.filter((eachlist) => currentPath.includes(eachlist.path));
+  const placeholders = await fetchPlaceholders();
+  const homePlaceholder = placeholders.breadcrumbsHomeLabel || 'Home';
+  breadcrumbsList.sort((a, b) => a.path.split('/').length - b.path.split('/').length).forEach((link) => {
+    if (link.path === '/') {
+      crumbs.push({ title: homePlaceholder, url: link.path });
+    } else {
+      crumbs.push({ title: link.breadcrumbstitle, url: link.path });
+    }
+  });
+
+  return crumbs;
+}
+
+async function buildBreadcrumbs() {
+  const breadcrumbs = document.createElement('nav');
+  breadcrumbs.className = 'breadcrumbs';
+
+  const crumbs = await buildBreadcrumbsFromNavTree(document.querySelector('.nav-sections'), document.location.href);
+
+  const ol = document.createElement('ol');
+  ol.append(...crumbs.map((item) => {
+    const li = document.createElement('li');
+    if (item['aria-current']) li.setAttribute('aria-current', item['aria-current']);
+    if (item.url) {
+      const a = document.createElement('a');
+      a.href = item.url;
+      a.textContent = item.title;
+      li.append(a);
+    } else {
+      li.textContent = item.title;
+    }
+    return li;
+  }));
+
+  breadcrumbs.append(ol);
+  return breadcrumbs;
+}
+
+export async function autoBlockBreadcrumb() {
+  if (getMetadata('breadcrumbs').toLowerCase() === 'true') {
+    const section = document.querySelector('main .section');
+    if (section && !section.dataset.breadcrumbsStatus) {
+      section.dataset.breadcrumbsStatus = 'initialized';
+      const block = await buildBreadcrumbs();
+      section.prepend(block);
+    }
+  }
 }
